+++ src/demo/Kinetic.tsx
@@ -1,0 +1,77 @@
+import type { SyntheticEvent } from "react";
+import React from "react";
+import { VanillaKinetic } from "..";
+
+// @ts-expect-error Idk why TS started complaining here on Next.js project.
+import styles from "./index.module.scss";
+
+interface Props {
+    onZoom: (zoomMultiplier: number, zoomLevel: number) => void;
+}
+
+class Kinetic extends React.Component<Props> {
+    public constructor(props: Props) {
+        super(props);
+    }
+
+    public shouldComponentUpdate() {
+        return false;
+    }
+
+    private _mapInstance?: VanillaKinetic | null;
+
+    private readonly onContainerRef = (ref: HTMLDivElement | null) => {
+        if (!ref) {
+            if (this._mapInstance) {
+                if (this._mapInstance.active) {
+                    this._mapInstance.destroy();
+                }
+                this._mapInstance = null;
+            }
+            return;
+        }
+        this._mapInstance = new VanillaKinetic(ref, {});
+        this._mapInstance.on("zoom", this.props.onZoom);
+        this._mapInstance.center();
+
+        const img = ref.querySelector("img");
+        if (img?.naturalWidth) {
+            this.handleImageLoad({ currentTarget: img });
+        }
+    };
+
+    private readonly handleImageLoad = (evt: Pick<SyntheticEvent<HTMLImageElement>, "currentTarget">) => {
+        const target = evt.currentTarget;
+        this._mapInstance?.setNaturalDimensions(target.naturalWidth, target.naturalHeight);
+        // this._mapInstance?.setZoom("fit");
+        this._mapInstance?.center();
+    };
+
+    private readonly _handleAttachDetach = () => {
+        if (!this._mapInstance) {
+            return;
+        }
+
+        if (this._mapInstance.active) {
+            this._mapInstance.destroy();
+        }
+        else {
+            this._mapInstance.reinitialize();
+        }
+    };
+
+    public render() {
+        return (
+            <>
+                <button onClick={this._handleAttachDetach}>attach / detach</button>
+                <div ref={this.onContainerRef} className={styles.container}>
+                    <div className={styles.middle}>
+                        <img src={"https://i.imgur.com/6UhKWdy.jpeg"} alt={"Wallpaper"} onLoad={this.handleImageLoad} />
+                    </div>
+                </div>
+            </>
+        );
+    }
+}
+
+export { Kinetic };