+++ src/types.ts
@@ -1,0 +1,7 @@
+interface KineticEvents {
+    zoom: [number, number];
+}
+
+export type {
+    KineticEvents,
+};